% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastLog.R
\name{beastLog}
\alias{beastLog}
\title{XML Parameter Nodes}
\usage{
beastLog(l, x, i)
}
\arguments{
\item{l}{A character string giving the name (i.e. \code{id} or \code{idref} attribute) of the desired parameter.}

\item{x}{A list, internally generated by \code{rbeauti}.}

\item{i}{Integer giving the index to a partition, internally generated by \code{rbeauti}.}
}
\value{
An object of class \code{XMLNode}.
}
\description{
A collection of parameter nodes used in the standard template of BEAST2.
}
\keyword{internal}
