% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastn.R
\name{blastn}
\alias{blastn}
\title{Nucleotide-Nucleotide BLAST}
\usage{
blastn(query, db)
}
\arguments{
\item{query}{An object of class \code{DNAbin} containing the sequences which
will be blasted against \code{db}.}

\item{db}{An object of class \code{DNAbin} containing the reference
sequences, i.e. the sequences against which \code{query} will be blasted.}
}
\description{
Provides an interface to BLASTN
}
