% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del.miss.R
\name{del.miss}
\alias{del.miss}
\title{Delete Missing Data from DNA Sequences}
\usage{
del.miss(x)
}
\arguments{
\item{x}{A matrix, a list, or a vector of class \code{\link[ape]{DNAbin}}
containing the DNA sequences.}
}
\value{
A list or a vector of class \code{DNAbin}.
}
\description{
Remove gaps ("-") and/or missing and ambiguous data ("N", "?") 
from a sample of DNA sequences.
}
