% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAbin2index.R
\name{DNAbin2index}
\alias{DNAbin2index}
\title{Conversion of DNAbin to Index}
\usage{
DNAbin2index(x)
}
\arguments{
\item{x}{A matrix of class \code{\link[ape]{DNAbin}}.}
}
\description{
Extract the indices of non-empty positions in a sample of DNA
  sequences to zzz.
}
\seealso{
\code{\link{index2DNAbin}}
}
