% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNodes.R
\name{fixNodes}
\alias{fixNodes}
\title{Standard Node Numbering in Phylo Objects}
\usage{
fixNodes(phy)
}
\arguments{
\item{phy}{An object of class \code{\link[ape]{phylo}}.}
}
\value{
An object of class \code{\link[ape]{phylo}}.
}
\description{
(Re-)establishes the standard numbering of terminal and internal
  nodes in phylogenies represented as objects of class \code{\link[ape]{phylo}}.
}
\details{
Internal and terminal nodes of \code{\link[ape]{phylo}} objects have a standard,
"canonical" node numbering. It consists of the following rules:
\enumerate{
 \item Number tip nodes from left/bottom to right/top by 1 to \code{ntip()}
 \item Number internal nodes by preorder traversal (see Examples) consecutively
 from \code{ntip() + 1} to \code{nnode() + 1}
}
 
When reading phylogenetic trees from a NEXUS file that contains a
\code{translate} section, it can happen that the terminal nodes (tips,
leaves) of the corresponding \code{phylo} object are not numbered
consecutively, which can be a problem in some downstream applications. You
can use \code{fixNodes} to get the correct order of terminal node numbers.

\code{fixNodes} is also intended to re-establish the standard numbering of
internal nodes and reorder all node value elements (e.g. node.label,
posterior, ...) if a \code{\link[ape]{phylo}} object has been modified by
either \code{\link[ape]{root}}, \code{\link[ape]{ladderize}}, or
\code{\link[ape]{rotate}}.
}
\examples{
set.seed(9999)

# Create random topology with random node numbering
phy1 <- rtopology(10)

# Impose 'canonical' node numbering
phy2 <- fixNodes(phy1)

# Compare node numbering schemes
op = par(no.readonly = TRUE)
par(mfcol = c(1, 2), mar = c(0, 0, 3, 0))
plot(phy1, use.edge.length = FALSE, show.tip.label = FALSE)
nodelabels(); tiplabels()
title("random node numbers")
plot(phy2, use.edge.length = FALSE, show.tip.label = FALSE)
nodelabels(); tiplabels()
title("preorder traversal of internal nodes")
par(op)

}
\seealso{
\code{\link[ape]{read.tree}}, \code{\link[ape]{read.nexus}},
  \code{\link{read.beast}} for reading trees in NEWICK and NEXUS format;
  \code{\link[ape]{ladderize}} and \code{\link[ape]{rotate}} for tree
  manipulation; \code{node.support} for plotting node support values has been
  moved to package \bold{viper}. %\code{\link{node.trans}} for handling
  non-standard node elements in class \code{\link[ape]{phylo}}.
}
