% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index2DNAbin.R
\name{index2DNAbin}
\alias{index2DNAbin}
\title{Conversion of Index to DNAbin}
\usage{
index2DNAbin(DNAbin, index)
}
\arguments{
\item{DNAbin}{A list of class \code{\link[ape]{DNAbin}}.}

\item{index}{A list of integers containing the indices of base positions.}
}
\description{
Use indices of non-empty positions to convert a list of 
DNA sequences into a matrix.
}
\seealso{
\code{\link{DNAbin2index}}
}
