\name{ips-internal} 
\alias{extractBEASTstats} % read.beast
\alias{extractMRBAYESstats} % read.mrbayes
\alias{phylo2mafft} % mafft
\alias{rbeauti.taxonset} % rbeauti
\alias{read.xml} % EXPERIMENTAL!
\alias{terminal.clades} % phylo2mafft
\alias{traitRateOutput} % traitRate
\alias{traitRateParams} % traitRate
\alias{unresolve.clade} % EXPERIMENTAL!

\title{Internal IPS Functions}
\description{Internal \pkg{ips} functions.}
\note{
These are internal functions in \bold{ips} and are not intended to be called by the user.
}
\keyword{internal}
