\name{ips.16S}
\alias{ips.16S}
\docType{data}
\title{Bark Beetle 16S Sequences}
\description{
This DNA alignment contains 376 positions of 42 sequences of 16S ribosomal DNA of the bark beetle genera \emph{Ips}, \emph{Orthotomicus}, and \emph{Pityogenes} (Scolytinae, Curculionidae, Coleoptera).
}
\usage{data(ips.16S)}
\format{
  The sequences are stored in binary format (see \code{\link[ape]{DNAbin}}).
}
\source{
The sequences were downloaded and assembled from the Nucleotide repository at GenBank on February 8, 2014.  
}
\references{
The nucleotide database on the NCBI website: \url{https://www.ncbi.nlm.nih.gov/nuccore}
}
\examples{
data(ips.16S)
}
\keyword{datasets}
