\name{ips.28S}
\alias{ips.28S}
\docType{data}
\title{
Bark Beetle 28S Sequences
}
\description{
This DNA alignment contains 562 positions of 28 sequences of 28S ribosomal DNA of the bark beetle genus \emph{Ips} (Scolytinae, Curculionidae, Coleoptera).
}
\usage{data(ips.28S)}
\format{
  The sequences are stored in binary format (see \code{\link[ape]{DNAbin}}).
}
\source{
The sequences were downloaded and assembled from the Nucleotide repository at GenBank on February 8, 2014. 
}
\references{
The nucleotide database on the NCBI website: \url{https://www.ncbi.nlm.nih.gov/nuccore}
}
\examples{
data(ips.28S)
}
\keyword{datasets}
