\name{ips.cox1}
\alias{ips.cox1}
\docType{data}
\title{
Bark Beetle COX1 Sequences
}
\description{
This DNA alignment contains 770 positions of 26 sequences of cox1 of the bark beetle genera \emph{Ips}, \emph{Orthotomicus}, and \emph{Pityogenes} (Scolytinae, Curculionidae, Coleoptera).
}
\usage{data(ips.cox1)}
\format{
  The sequences are stored in binary format (see \code{\link{DNAbin}}).
}
\source{
The sequences were downloaded and assembled from the Nucleotide resitory at GenBank on February 8, 2014. 
}
\references{
The nucleotide database on the NCBI website: \url{http://www.ncbi.nlm.nih.gov/nuccore}
}
\examples{
data(ips.cox1)
}
\keyword{datasets}
