% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIPSproj.R
\name{LIPS_proj}
\alias{LIPS_proj}
\title{'Local' Integrated Propensity Score estimator based on projection weighting function}
\usage{
LIPS_proj(
  z,
  d,
  x,
  xbal = NULL,
  beta.initial = NULL,
  lin.rep = TRUE,
  whs = NULL,
  x_keep = FALSE,
  maxit = 50000,
  allRows = FALSE,
  kernel = NULL,
  init.method = c("glm", "CBPS")
)
}
\arguments{
\item{z}{An \eqn{n} x \eqn{1} vector of binary instruments.}

\item{d}{An \eqn{n} x \eqn{1} vector of binary treatment adoption indicators.}

\item{x}{An \eqn{n} x \eqn{k}  matrix of covariates to be used in the propensity score. First element must be a vector of 1's.}

\item{xbal}{An \eqn{n} x \eqn{l}, \eqn{l\leq k}, matrix of ``raw'' covariares to be balanced (does not need to include interaction terms). Default is \code{NULL}, which will use the same as x.}

\item{beta.initial}{An optional \eqn{k} x \eqn{1} vector of initial values for the parameters to be optimized over.}

\item{lin.rep}{Logical argument to whether an estimator for the asymptotic linear representation of the LIPS
parameters should be provided. Deafault is TRUE.}

\item{whs}{An optional \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights.}

\item{x_keep}{Default is FALSE. If TRUE, we return covariate matrix in the output.}

\item{maxit}{The maximum number of iterations. Defaults to 50000.  = FALSE). Deafault is 999 if boot = TRUE}

\item{allRows}{Retained for backward compatibility. Duplicate rows are now handled automatically and do not need to be sorted.}

\item{kernel}{Optional precomputed IPS kernel from \code{\link{ips_kernel}}.
If supplied, \code{xbal} is not used to build a new kernel.}

\item{init.method}{Method used to compute starting values when
\code{beta.initial = NULL}. Default is \code{"glm"}; \code{"CBPS"} uses the
optional \pkg{CBPS} package.}
}
\value{
A list containing the following components:
\item{coefficients}{The estimated LIPS_proj coefficients}
\item{fitted.values}{The LIPS_proj fitted probabilities}
\item{linear.predictors}{The LIPS_proj estimated index (X'beta)}
\item{lin.rep}{An estimator of the LIPS_proj coefficients' asymptotic linear representation}
\item{converged}{An integer code. 0 indicates successful completion}
\item{x}{The model matrix (i.e. the matrix of covariates used to estimate the LIPS_proj parameters). Only returned if \code{x_keep = TRUE}.}
}
\description{
'Local' Integrated Propensity Score estimator based on projection weighting function
}
\references{
Sant'Anna, Pedro H. C., Song, Xiaojun, and Xu, Qi (2022), \emph{Covariate distribution balance via propensity scores},
      \emph{Journal of Applied Econometrics}, 37(6), 1093-1120. <https://doi.org/10.1002/jae.2909>.
}
