% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbayes.lset.R
\name{mrbayes.lset}
\alias{mrbayes.lset}
\title{MrBayes model settings}
\usage{
mrbayes.lset(
  nucmodel = "4by4",
  nst = 1,
  code = "universal",
  ploidy = "diploid",
  rates = "equal",
  ngammacat = 4,
  nlnormcat = 4,
  nmixtcat = 4,
  nbetacat = 5,
  omegavar = "equal",
  covarion = "no",
  coding = "all",
  parsmodel = "no",
  partition = 1
)
}
\arguments{
\item{nucmodel}{A character string}

\item{nst}{xxx}

\item{code}{xxx}

\item{ploidy}{xxx}

\item{rates}{xxx}

\item{ngammacat}{xxx}

\item{nlnormcat}{xxx}

\item{nmixtcat}{xxx}

\item{nbetacat}{xxx}

\item{omegavar}{xxx}

\item{covarion}{xxx}

\item{coding}{xxx}

\item{parsmodel}{xxx}

\item{partition}{A character string serving as an identifier in a
multi-partition model.}
}
\value{
A list of model parameters.
}
\description{
Specify a model of DNA, Protein or trait evolution for \code{\link{mrbayes}}.
}
\details{
xxx
}
\seealso{
\code{\link{mrbayes}}, \code{\link{mrbayes.prset}} , \code{\link{mrbayes.mcmc}}.
}
