% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbayes.mcmc.R
\name{mrbayes.mcmc}
\alias{mrbayes.mcmc}
\title{MrBayes MCMC settings}
\usage{
mrbayes.mcmc(
  ngen = 1e+06,
  nruns = 2,
  nchains = 4,
  temp = 0.1,
  reweight = c(0, 0),
  swapfreq = 1,
  nswaps = 1,
  samplefreq = 500,
  printfreq = 1000,
  printall = TRUE,
  printmax = 8,
  mcmcdiagn = TRUE,
  diagnfreq = 5000,
  diagnstat = "avgstddev",
  minpartfreq = 0.1,
  allchains = FALSE,
  allcomps = FALSE,
  relburnin = TRUE,
  burnin = 0,
  burninfrac = 0.25,
  stoprule = FALSE,
  stopval = 0.05,
  savetrees = FALSE,
  checkpoint = TRUE,
  checkfreq = 2000,
  startparams = "current",
  starttree = "current",
  nperts = 0,
  data = TRUE,
  ordertaxa = FALSE,
  append = FALSE,
  autotune = TRUE,
  tunefreq = 100
)
}
\arguments{
\item{ngen}{A character string}

\item{nruns}{xxx}

\item{nchains}{xxx}

\item{temp}{xxx}

\item{reweight}{xxx}

\item{swapfreq}{xxx}

\item{nswaps}{xxx}

\item{samplefreq}{xxx}

\item{printfreq}{xxx}

\item{printall}{xxx}

\item{printmax}{xxx}

\item{mcmcdiagn}{xxx}

\item{diagnfreq}{xx}

\item{diagnstat}{xxx}

\item{minpartfreq}{xxx}

\item{allchains}{xxx}

\item{allcomps}{xxx}

\item{relburnin}{xxx}

\item{burnin}{xxx}

\item{burninfrac}{xxx}

\item{stoprule}{xxx}

\item{stopval}{xxx}

\item{savetrees}{xxx}

\item{checkpoint}{xxx}

\item{checkfreq}{xxx}

\item{startparams}{xxx}

\item{starttree}{xxx}

\item{nperts}{xxx}

\item{data}{xxx}

\item{ordertaxa}{xxx}

\item{append}{xxx}

\item{autotune}{xxx}

\item{tunefreq}{xxx}
}
\value{
A list of MCMC parameters.
}
\description{
Specify the MCMC settings for \code{\link{mrbayes}}.
}
\details{
xxx
}
\seealso{
\code{\link{mrbayes}}, \code{\link{mrbayes.lset}} , \code{\link{mrbayes.prset}}.
}
