% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbayes.prset.R
\name{mrbayes.prset}
\alias{mrbayes.prset}
\title{MrBayes prior settings}
\usage{
mrbayes.prset(
  traitiopr,
  revmatpr,
  aamodelpr,
  aarevmatpr,
  omegapr,
  ny98omega1pr,
  ny98omega3pr,
  m3omegapr,
  codoncatfreqs,
  statefreqpr,
  shapepr,
  ratecorrpr,
  pinvarpr,
  covswitchpr,
  symdirihyperpr,
  topologypr,
  brlenspr,
  treeagepr,
  speciationpr,
  extinctionpr,
  fossilizationpr,
  samplestrat,
  sampleprop,
  popsizepr,
  popvarpr,
  nodeagepr,
  clockratepr,
  clockvarpr,
  cppratepr,
  cppmultdevpr,
  tk02varpr,
  wnvarpr,
  igrvarpr,
  ilnvarpr,
  ratepr,
  generatepr
)
}
\arguments{
\item{traitiopr}{A character string}

\item{revmatpr}{xxx}

\item{aamodelpr}{xxx}

\item{aarevmatpr}{xxx}

\item{omegapr}{xxx}

\item{ny98omega1pr}{xxx}

\item{ny98omega3pr}{xxx}

\item{m3omegapr}{xxx}

\item{codoncatfreqs}{xxx}

\item{statefreqpr}{xxx}

\item{shapepr}{xxx}

\item{ratecorrpr}{xxx}

\item{pinvarpr}{xxx}

\item{covswitchpr}{xxx}

\item{symdirihyperpr}{xxx}

\item{topologypr}{xxx}

\item{brlenspr}{xxx}

\item{treeagepr}{xxx}

\item{speciationpr}{xxx}

\item{extinctionpr}{xxx}

\item{fossilizationpr}{xxx}

\item{samplestrat}{xxx}

\item{sampleprop}{xxx}

\item{popsizepr}{xxx}

\item{popvarpr}{xxx}

\item{nodeagepr}{xxx}

\item{clockratepr}{xxx}

\item{clockvarpr}{xxx}

\item{cppratepr}{xxx}

\item{cppmultdevpr}{xxx}

\item{tk02varpr}{xxx}

\item{wnvarpr}{xxx}

\item{igrvarpr}{xxx}

\item{ilnvarpr}{xxx}

\item{ratepr}{xxx}

\item{generatepr}{xxx}
}
\value{
A list of parameters for prior distributions.
}
\description{
Specify parameters for prior distributions for \code{\link{mrbayes}}.
}
\details{
xxx
}
\seealso{
\code{\link{mrbayes}}, \code{\link{mrbayes.lset}} , \code{\link{mrbayes.mcmc}}.
}
