% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighboringPairs.R
\name{neighboringPairs}
\alias{neighboringPairs}
\title{Neighboring Nodes in a Minimum Spanning Tree}
\usage{
neighboringPairs(mst)
}
\arguments{
\item{mst}{An object of class \code{\link{mst}}.}
}
\description{
Finds all pairs of adjacent nodes, i.e. 
nodes separated by only one edge, in a minimum spanning 
tree
}
