% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntip.R
\name{ntip}
\alias{ntip}
\title{Numbers of Tips of (Sub)trees}
\usage{
ntip(phy, node)
}
\arguments{
\item{phy}{An object of class \code{\link[ape:read.tree]{phylo}}.}

\item{node}{An integer given the number of an internal node.}
}
\value{
An integer giving the number of tips.
}
\description{
Counts the number of tips of a given clade of a 
phylogenetic tree.
}
\examples{
set.seed(1234)
tr <- rtree(12)
plot(tr); nodelabels()
ntip(tr, 16)
}
