% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionfinder.R
\name{partitionfinder}
\alias{partitionfinder}
\title{PartitionFinder}
\usage{
partitionfinder(
  alignment,
  user.tree,
  branchlengths = "linked",
  models = "all",
  model.selection = "BIC",
  search = "greedy",
  exec = "/Applications/PartitionFinderV1.1.1_Mac/PartitionFinder.py"
)
}
\arguments{
\item{alignment}{A}

\item{user.tree}{A}

\item{branchlengths}{A}

\item{models}{A}

\item{model.selection}{A}

\item{search}{A}

\item{exec}{A character string giving the path to the executable (python 
script).}
}
\description{
Provides a wrapper to the PartitionFinder software.
}
\references{
Lanfear, R., B. Calcott, S.Y.W. Ho, and S. Guindon. 2012. 
  PartitionFinder: combined selection of partitioning schemes and
  substitution models for phylogenetic analyses. \emph{Molecular 
  Biology and Evolution} \strong{29}: 1695-1701.  
  
  Lanfear, R., B. Calcott, K. David, C. Mayer, and A. Stamatakis. 2014. 
  Selecting optimal partitioning schemes for phylogenomic datasets.
  \emph{BMC Evolutionary Biology} \strong{14}: 82.
}
