% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo2mst.R
\name{phylo2mst}
\alias{phylo2mst}
\title{Conversion from PHYLO to MST Object}
\usage{
phylo2mst(phy)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}
}
\description{
Converts a phylogenetic tree (class \code{phylo}) into a
  minimum spanning tree (class \code{mst}).
}
\details{
The current version of \code{phylo2mst} does not handle polytomies
  and does not incorporate branch length information. Note that topological 
  information is lost during the conversion.
}
\examples{
phy <- rtree(12)
plot(phy)
mst <- phylo2mst(phy)
plot(mst)
}
