% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTT.R
\name{QTT}
\alias{QTT}
\title{IPW estimator for the Quantile Treatment Effect on the Treated}
\usage{
QTT(y, d, x, ps, beta.lin.rep, tau = 0.5, bw = "nrd0", trim = FALSE,
  trim.at = NULL, whs = NULL)
}
\arguments{
\item{y}{An \eqn{n} x \eqn{1} vector of outcome of interest.}

\item{d}{An \eqn{n} x \eqn{1} vector of binary treatment adoption indicators.}

\item{x}{An \eqn{n} x \eqn{k}  matrix of covariates used in the propensity score estimation}

\item{ps}{An \eqn{n} x \eqn{1} vector of fitted propensity scores.}

\item{beta.lin.rep}{An \eqn{n} x \eqn{k}  matrix of estimates of the asymptotic linear representaion of the propensity score parameters (used to compute std. errors).}

\item{tau}{An \eqn{l} x \eqn{1} vector of quantile to compute the QTT at. If NULL, then we set tau = 0.5.}

\item{bw}{Bandwidth choice to compute densities (used to compute std. errors). Options are "ucv","nrd", "nrd0", "bcv", "SJ" - see bw.nrd for additional details.
Default choice is "nrd0".}

\item{trim}{Logical argument to whether one should trim propensity scores. Deafault is FALSE.}

\item{trim.at}{Only used if trim=TRUE. If a scalar, trim all propensity score below trim.at and above 1 - trim.at.
If a  \eqn{2} x \eqn{1} vector, trim all propensity scores below trim.at[1] and all propensity scores above trim.at[2].
If NULL, trim.at is set to 1e-10.}

\item{whs}{An optional \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights.}
}
\value{
A list containing the following components:
\item{qtt}{The estimated QTT}
\item{qtt.se}{Estimated (pointwise) std. error of the QTT.}
\item{qtt.inf}{Estimated influence function of QTT estimator.}
\item{tau}{The evaluation points of QTT.}
}
\description{
IPW estimator for the Quantile Treatment Effect on the Treated
}
\references{
Sant'Anna, Pedro H. C, Song, Xiaojun, and Xu, Qi (2019), \emph{Covariate Distribution Balance via Propensity Scores},
      Working Paper <https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3258551>.
}
