% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootedPASTA.R
\name{rootedPASTA}
\alias{rootedPASTA}
\title{PASTA with a Rooted Guide Tree}
\usage{
rootedPASTA(phy, k)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{k}{An integer giving the maximum size of the subgroups that will be 
aligned with MAFFT.}
}
\description{
Implements the PASTA (ultra-large multiple sequence alignment) 
algorithm of Mirarab, Nguyen, and Warnow (2014) with a rooted guide tree.
}
\references{
Mirarab, S., N. Nguyen, and T. Warnow. 2014. PASTA: Ultra-large multiple
sequence alignment. \emph{RECOMB} \strong{}: 177-191.
}
