% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitEqualSubtrees.R
\name{splitEqualSubtrees}
\alias{splitEqualSubtrees}
\title{Split Phylogenetic Trees}
\usage{
splitEqualSubtrees(phy, k)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{k}{An integer, only trees with > k tips will be split.}
}
\value{
A list of two phylogenetic trees.
}
\description{
Splits a phylogenetic tree into roughly equally sized
subtrees
}
\examples{
set.seed(1234)
tr <- rtree(12000)
subtr <- splitEqualSubtrees(tr)
subtr
}
\seealso{
\code{\link{extract.clade}}.
}

