% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitIntoClades.R
\name{splitIntoClades}
\alias{splitIntoClades}
\title{Find Monophyletic Subsets in Species Lists}
\usage{
splitIntoClades(phy, tips)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{tips}{A vector of mode \code{"character"} containing any 
subset of the tiplabels in \code{phy}.}
}
\value{
A list.
}
\description{
Takes a phylogeny and a subset of its tiplabels
and splits the list of tiplabels into monophyletic groups 
(clades).
}
