\name{ips-internal} 
\alias{are.tips.consecutive} % noi
\alias{assembleDataNode} % rbeauti
\alias{assembleDistributionNode} % rbeauti
\alias{assembleInitNode} % rbeauti
\alias{assembleLoggers} % rbeauti
\alias{assembleOperators} % rbeauti
\alias{assembleStateNode} % rbeauti
\alias{assembleTreeNode} % rbeauti
\alias{extractBEASTstats} % read.beast
\alias{extractMRBAYESstats} % read.mrbayes
\alias{formatSet} % mrbayes
\alias{log_list} % assembleLoggers
\alias{matrixBlock} % write.nex
\alias{operator} % rbeauti
\alias{operator_list} % rbeauti
\alias{parameter} % several functions
\alias{rbeauti.taxonset} % rbeauti
\alias{rc.core}
\alias{read.xml} % EXPERIMENTAL!
\alias{setClock} % rbeauti
\alias{setMCMC} % rbeauti
\alias{str2cha} % write.phy, write.nex
\alias{terminal.clades} % phylo2mafft
\alias{traitRateOutput} % traitRate
\alias{traitRateParams} % traitRate
\alias{unresolve.clade} % EXPERIMENTAL!

\title{Internal IPS Functions}
\description{Internal \pkg{ips} functions.}
\note{
These are internal functions in \bold{ips} and are not intended to be called by the user.
}
\keyword{internal}
