% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIPSproj.R
\name{LIPS_proj}
\alias{LIPS_proj}
\title{'Local' Integrated Propensity Score estimator based on projection weighting function}
\usage{
LIPS_proj(
  z,
  d,
  x,
  beta.initial = NULL,
  lin.rep = TRUE,
  whs = NULL,
  maxit = 50000,
  allRows = F
)
}
\arguments{
\item{z}{An \eqn{n} x \eqn{1} vector of binary instruments.}

\item{d}{An \eqn{n} x \eqn{1} vector of binary treatment adoption indicators.}

\item{x}{An \eqn{n} x \eqn{k}  matrix of covariates to be used in the propensity score. First element must be a vector of 1's.}

\item{beta.initial}{An optional \eqn{k} x \eqn{1} vector of initial values for the parameters to be optimized over.}

\item{lin.rep}{Logical argument to whether an estimator for the asymptotic linear representation of the LIPS
parameters should be provided. Deafault is TRUE.}

\item{whs}{An optional \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights.}

\item{maxit}{The maximum number of iterations. Defaults to 50000.  = FALSE). Deafault is 999 if boot = TRUE}

\item{allRows}{Default is FALSE, which attempts to fist check if all rows of the matrix x are unique. If there are draws, it tries to optimize the code, but requires 'x' to be sorted such that all unique rows are together.}
}
\value{
A list containing the following components:
\item{coefficients}{The estimated LIPS_proj coefficients}
\item{fitted.values}{The LIPS_proj fitted probabilities}
\item{linear.predictors}{The LIPS_proj estimated index (X'beta)}
\item{lin.rep}{An estimator of the LIPS_proj coefficients' asymptotic linear representation}
\item{converged}{An integer code. 0 indicates successful completion}
\item{x}{The model matrix (i.e. the matrix of covariates used to estimate the LIPS_proj parameters)}
}
\description{
'Local' Integrated Propensity Score estimator based on projection weighting function
}
\references{
Sant'Anna, Pedro H. C, Song, Xiaojun, and Xu, Qi (2019), \emph{Covariate Distribution Balance via Propensity Scores},
      Working Paper <https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3258551>.
}
