% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalSisters.R
\name{terminalSisters}
\alias{terminalSisters}
\title{Find Pairs of Sister Species}
\usage{
terminalSisters(phy, labels = TRUE)
}
\arguments{
\item{phy}{An object of class \code{\link{phylo}}.}

\item{labels}{Logical, indicating whether to return tip labels or tip numbers.}
}
\value{
A list of which each element contains the tip labels
of a sister species pair.
}
\description{
Finds pairs of sister species in
a phylogenetic tree.
}
\examples{
set.seed(1234)
tr <- rtree(12)
plot(tr)
terminalSisters(tr)
}
